+++ CHANGELOG.md
@@ -8,6 +8,9 @@ Changelog
 * Set bugsnag.span.category to 'custom' for custom spans.
   [336](https://github.com/bugsnag/bugsnag-cocoa-performance/pull/336)
 
+* Fixed a crash after shared NSURLSession invalidate
+  [334](https://github.com/bugsnag/bugsnag-cocoa-performance/pull/334)
+
 * Fix visionOS compilation errors. Note that visionOS is not yet officially supported.
   [327](https://github.com/bugsnag/bugsnag-cocoa-performance/pull/327)